requirements.txt
asgiref==3.5.2
Django==4.0.4
django-tinymce==3.4.0
et-xmlfile==1.1.0
lml==0.1.0
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
psycopg2==2.9.3
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
